+++ js/util.js
@@ -1,0 +1,42 @@
+const getRandomInteger = function (min, max) {
+  if (min < 0 || max < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const checkStringLangth = (string, maxLength) => string.length <= maxLength;
+checkStringLangth('1sting', 11);
+
+const createRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+function createIdGenerator () {
+  let lastGeneratedId = 0;
+
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+}
+
+const createArrayElement = (elements) => (
+  elements[getRandomInteger(0, elements.length - 1)]
+);
+
+export {getRandomInteger, createRandomIdFromRangeGenerator, createIdGenerator, createArrayElement};
